+++ falco/__init__.py
@@ -2,15 +2,16 @@
 from falco.config import init_from_mat as ifm
 from .setup import *
 from .utils import *
-from .wfsc  import *
 from .configs  import *
 from .imaging import *
 from .dms import *
-from .setups import *
+from .est  import *
+from .ctrl  import *
+from .wfsc  import *
 from .prop import *
+from .segmentutils import *
 from .zernikes import *
 from . import model
 from ._globals import INFLUENCE_XINETICS
 from ._globals import INFLUENCE_BMC_KILO
 from ._globals import INFLUENCE_BMC_2K
-from .segmentutils import *
