--- falco/setups.py
@@ -1,42 +1,0 @@
-def falco_setup_FPM_HLC(mp):
-    """
-    Quick Description here
-
-    Detailed description here
-
-    Parameters
-    ----------
-    mp: falco.config.ModelParameters
-        Structure of model parameters
-    Returns
-    -------
-    TBD
-        Return value descriptio here
-    """
-
-    if type(mp) is not falco.config.ModelParameters:
-        raise TypeError('Input "mp" must be of type ModelParameters')
-    pass
-
-def falco_setup_FPM_HLC_cosine(mp):
-    """
-    Quick Description here
-
-    Detailed description here
-
-    Parameters
-    ----------
-    mp: falco.config.ModelParameters
-        Structure of model parameters
-    Returns
-    -------
-    TBD
-        Return value descriptio here
-    """
-
-    if type(mp) is not falco.config.ModelParameters:
-        raise TypeError('Input "mp" must be of type ModelParameters')
-    pass
-
-
-
